Update README.md
When naming your ATM project in the backend code, you can choose a name that reflects the purpose and functionality of the project. For example, you could use something like "ATMSystem," "CashMachine," or "ATMService" to indicate that the code is related to the ATM project. The key is to use a name that is descriptive and easy to understand for anyone working with the code.